requirements.txt
Flask==1.1.2
flask_uploads
flask_dropzone
h5py==2.10.0
imageio==2.8.0
Jinja2==2.11.2
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
matplotlib==3.2.1
mmcv==0.4.4
numpy==1.18.3
opencv-contrib-python==4.2.0.34
opencv-python==4.2.0.34
pandas==1.0.3
Pillow==7.1.1
protobuf==3.11.3
Cython
pycocotools==2.0.0
scikit-image==0.16.2
scikit-learn==0.22.2.post1
scipy==1.4.1
tensorboard==1.13.1
tensorflow==1.13.2
tensorflow-estimator==1.13.0
tensorpack==0.10
torch==1.5.0
torchvision==0.6.0
redis
